Library zim file detail controller
